@@ -6,44 +6,44 @@
     "dev": "next dev",
     "build": "next build",
     "start": "next start",
-    "sitemap": "node scripts/generate-sitemap.js",
-    "feed": "node ./.next/server/scripts/generate-feed.js",
-    "export": "npm run build && next export && npm run sitemap && npm run feed"
+    "sitemap": "node scripts/generate-sitemap.mjs",
+    "export": "npm run build && next export && npm run sitemap"
   },
   "dependencies": {
-    "@fortawesome/fontawesome-free": "^5.15.4",
-    "@fortawesome/fontawesome-svg-core": "~1.2.36",
-    "@fortawesome/free-brands-svg-icons": "^5.15.4",
-    "@fortawesome/free-regular-svg-icons": "^5.15.4",
-    "@fortawesome/free-solid-svg-icons": "^5.15.4",
-    "@fortawesome/react-fontawesome": "^0.1.17",
-    "@next/mdx": "^10.0.4",
+    "@fortawesome/fontawesome-free": "^6.2.1",
+    "@fortawesome/fontawesome-svg-core": "^6.2.1",
+    "@fortawesome/free-brands-svg-icons": "^6.2.1",
+    "@fortawesome/free-regular-svg-icons": "^6.2.1",
+    "@fortawesome/free-solid-svg-icons": "^6.2.1",
+    "@fortawesome/react-fontawesome": "^0.2.0",
+    "@mapbox/rehype-prism": "^0.8.0",
+    "@mdx-js/react": "^2.1.3",
+    "@next/mdx": "^13.0.6",
     "bootstrap": "^3.4.1",
-    "date-fns": "^2.16.1",
-    "disqus-react": "^1.0.11",
-    "feed": "^4.2.1",
+    "date-fns": "^2.29.3",
+    "disqus-react": "^1.1.5",
+    "feed": "^4.2.2",
     "lunr": "^2.3.9",
-    "next": "^10.0.0",
+    "next": "^13.0.6",
     "next-compose-plugins": "^2.2.1",
     "next-fonts": "^1.5.1",
     "next-ga": "^2.3.4",
     "ramda": "^0.27.1",
-    "react": "^17.0.1",
-    "react-dom": "^17.0.1",
-    "react-external-link": "^1.2.0",
-    "react-highlight": "^0.14.0",
-    "react-image-lightbox": "^5.1.1",
-    "react-tweet-embed": "^1.2.2"
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-external-link": "^2.0.0",
+    "react-image-lightbox": "^5.1.4",
+    "react-tweet-embed": "^2.0.0",
+    "rehype-infer-description-meta": "^1.1.0"
   },
   "devDependencies": {
-    "@mdx-js/loader": "^1.6.22",
-    "@types/lunr": "^2.3.3",
-    "@types/node": "^14.14.16",
+    "@mdx-js/loader": "^2.1.5",
+    "@types/lunr": "^2.3.4",
+    "@types/node": "^18.11.11",
     "@types/ramda": "^0.27.34",
-    "@types/react": "17.0.30",
-    "nextjs-sitemap-generator": "^1.1.3",
-    "optimize-css-assets-webpack-plugin": "^5.0.4",
-    "react-lunr": "^1.0.0",
-    "typescript": "^4.1.3"
+    "@types/react": "18.0.26",
+    "nextjs-sitemap-generator": "^1.3.1",
+    "react-lunr": "^1.1.0",
+    "typescript": "^4.9.4"
   }
 }
