@@ -1,9 +1,9 @@
-module.exports = {
+export const config = {
   GA: 'UA-38351554-3',
   DISQUS_SHORTNAME: 'acelayablog',
   CARBON_SERVE: 'CK7IVK3W',
   CARBON_PLACEMENT: 'blogalejandrocelayacom',
   SITE_URL: 'https://alejandrocelaya.blog',
   SITE_TITLE: 'Alejandro Celaya | Blog',
-  SITE_SUBTITLE: 'Software development, agile methodologies and open source projects'
+  SITE_SUBTITLE: 'Software development, agile methodologies and open source software'
 };
