@@ -1,6 +1,8 @@
-import { FC } from 'react';
+import { FC, PropsWithChildren } from 'react';
 
-export const Container: FC<{ className?: string }> = ({ children, className }) => (
+type ContainerProps = PropsWithChildren<{ className?: string }>;
+
+export const Container: FC<ContainerProps> = ({ children, className }) => (
   <div id="fh5co-faqs">
     <div className={`container ${className ?? ''}`}>
       {children}
