@@ -1,10 +1,9 @@
-import React, { FC, ReactNode } from 'react';
+import React, { FC, PropsWithChildren } from 'react';
 import { ExternalLink } from 'react-external-link';
 import InternalLink from './InternalLink';
 
-export interface LinkProps {
+export interface LinkProps extends PropsWithChildren {
   href: string;
-  children?: ReactNode;
   [rest: string]: any;
 }
 
