@@ -1,6 +1,6 @@
-import { FC } from 'react';
+import { FC, PropsWithChildren } from 'react';
 
-export const SecondaryContainer: FC = ({ children }) => (
+export const SecondaryContainer: FC<PropsWithChildren> = ({ children }) => (
   <div id="fh5co-subscribe">
     <div className="container">
       {children}
