@@ -1,15 +1,16 @@
-import { MDXProvider } from '@mdx-js/react';
 import { renderToStaticMarkup } from 'react-dom/server';
 import { mdxComponents } from './mdx';
 import { PostMeta } from './posts';
 
-export const renderPost = async (fileName: string): Promise<{ content: string; metadata: PostMeta }> => {
-  const { default: Content, metadata = {} } = await import(`../posts/${fileName}`);
-  const content = renderToStaticMarkup(
-    <MDXProvider components={mdxComponents}>
-      <Content />
-    </MDXProvider>
-  )
+interface RenderPostResult {
+  content: string;
+  metadata: PostMeta;
+  summary: string;
+}
 
-  return { content, metadata };
+export const renderPost = async (fileName: string): Promise<RenderPostResult> => {
+  const { default: Content, metadata = {}, summary = '' } = await import(`../posts/${fileName}`);
+  const content = renderToStaticMarkup(<Content components={mdxComponents}/>);
+
+  return { content, metadata, summary };
 }
